fix merge conflict when merge development to refactor-repositories-and-packaging
